\name{KScams_cov}
\alias{KScams_cov}
\docType{data}
\title{
KScams_cov
}
\description{
Habitat variables for KScams_dat
}
\usage{data("KScams_cov")}
\format{
  A data frame with 1149 observations on the following 10 variables.
  \describe{
    \item{\code{SurveyYear}}{Year of survey (a numeric vector)}
    \item{\code{Site}}{Site number (a numeric vector)}
    \item{\code{Latitude}}{Site latitude (a numeric vector)}
    \item{\code{Longitude}}{Site longitude (a numeric vector)}
    \item{\code{VegHeight}}{Height of vegetation at site(a numeric vector)}
    \item{\code{GrassPct}}{Percentage of grass at site (a numeric vector)}
    \item{\code{ForbPct}}{Percentage of forbs at site (a numeric vector)}
    \item{\code{CropPct}}{Percentage of crops at site (a numeric vector)}
    \item{\code{OpenGroundPct}}{Percentage of open ground at site (a numeric vector)}
    \item{\code{ShrubPct}}{Percentage of shrubs at site (a numeric vector)}
  }
}

\source{
URL: https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/ecy.3945
}
\references{
Palomo‐Munoz, G., T. J. Werdel, C. W. Piper, M. S. Peek, A. M. Ricketts, and A. A. Ahlers. 2023. Spatiotemporal distributions of mammals occurring in an agro‐prairie ecosystem. Ecology 104:e3945
}
\examples{
data(KScams_cov)
}
\keyword{datasets}
