\name{KScams_doy}
\alias{KScams_doy}
\docType{data}
\title{
KScams_doy
}
\description{
Days since 1 Jan of deployment year for KScams_dat
}
\usage{data("KScams_doy")}
\format{
  A data frame with 1149 observations including the following 36 variables.
  \describe{
    \item{\code{General_project}}{Project name (a character vector)}
    \item{\code{Year}}{Year of deployment (a numeric vector)}
    \item{\code{Site}}{Site number (a numeric vector)}
    \item{\code{Latitude}}{Site latitude in decimal degrees (a numeric vector)}
    \item{\code{Longitude}}{Site longitude in decimal degrees (a numeric vector)}
    \item{\code{Setup_date}}{Date of deployment (a character vector)}
    \item{\code{Retrieval_date}}{Date of retrieval (a character vector)}
    \item{\code{DaysActive}}{How many days camera was functioning (a numeric vector)}
    \item{\code{DOY_1}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_2}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_3}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_4}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_5}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_6}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_7}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_8}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_9}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_10}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_11}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_12}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_13}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_14}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_15}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_16}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_17}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_18}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_19}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_20}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_21}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_22}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_23}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_24}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_25}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_26}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_27}}{Day since 1 Jan of deployment year (a numeric vector)}
    \item{\code{DOY_28}}{Day since 1 Jan of deployment year (a numeric vector)}
  }
}
\source{
URL: https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/ecy.3945
}
\references{
Palomo‐Munoz, G., T. J. Werdel, C. W. Piper, M. S. Peek, A. M. Ricketts, and A. A. Ahlers. 2023. Spatiotemporal distributions of mammals occurring in an agro‐prairie ecosystem. Ecology 104:e3945
}
\examples{
data(KScams_doy)
}
\keyword{datasets}
