% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T3.R
\name{T3}
\alias{T3}
\title{Calculating the time between detections of species A to species A}
\source{
URL: \url{https://doi.org/10.1016/j.biocon.2016.09.001}
}
\usage{
T3(
  data,
  speciesA,
  speciesB,
  species_col,
  datetime_col,
  site_col,
  unitTime = "hours"
)
}
\arguments{
\item{data}{The camera trapping dataset containing information on species, datetime, and site (dataframe).}

\item{speciesA}{The first species in the interaction sequence (character).}

\item{speciesB}{The second species in the interaction sequence (character).}

\item{species_col}{The column name indicating the species in the dataset (character).}

\item{datetime_col}{The column name indicating the datetime of each detection (character).}

\item{site_col}{The column name indicating the camera site (character).}

\item{unitTime}{The unit of time used for calculating the time differences, default is "hours".
Options: "secs", "mins", "hours" (character).}
}
\value{
A list containing:
  \describe{
    \item{total_summary}{A summary of the mean values for T3 across all sites that recorded an event and years.}
    \item{event_count}{The total count of T3 events across all sites and years.}
    \item{event_summary}{The min1st & 3rd quartiles, median, mean, max for T3 events.}
    \item{site_summary}{A summary of the mean T3 for each site across all years, if no event is recorded NA will be reported.}
    \item{detailed_summary}{Detailed information on recorded T3 events, "including site, year and time" differences.}
  }
}
\description{
The T3 (AA) function analyzes camera trapping data to calculate the time between the detections of species A followed by species A without species B in between detections at each camera site for all years. The function returns the average of all T3 events for all sites and across all years, the number of times the event occurred, a site summary across all years and a detailed summary of each time the event occurred.
}
\examples{
# Function Example. Note: DateTime column must be formatted as a date-time and not a character string!
T3_example <- T3(data = KScams_dat, speciesA = "White-Tailed Deer", speciesB = "Coyote",
                  species_col = "Common_name", datetime_col = "DateTime",
                  site_col = "Site", unitTime = "hours")

}
\references{
Parsons, A. W., C. Bland, T. Forrester, M. C. Baker-Whatton, S. G. Schuttler, W. J. McShea, R. Costello, and R. Kays. 2016. The ecological impact of humans and dogs on wildlife in protected areas in eastern North America. Biological Conservation 203:75–88.
}
\seealso{
Niedballa, J., A. Wilting, R. Sollmann, H. Hofer, and A. Courtiol. 2019. Assessing analytical methods for detecting spatiotemporal interactions between species from camera trapping data. M. Rowcliffe and J. Ahumada, editors. Remote Sensing in Ecology and Conservation 5:272–285.

URL: https://doi.org/10.1002/rse2.107
}
\keyword{internal}
