\name{KScams_doy}
\alias{KScams_doy}
\docType{data}
\title{
KScams_doy
}
\description{
Dates of deployment (julian) for KScams_dat
}
\usage{data("KScams_doy")}
\format{
  A data frame with 1149 observations on the following 36 variables.
  \describe{
    \item{\code{General_project}}{Project name (a character vector)}
    \item{\code{Year}}{Year of deployment (a numeric vector)}
    \item{\code{Site}}{Site number (a numeric vector)}
    \item{\code{Latitude}}{Site latitude (a numeric vector)}
    \item{\code{Longitude}}{Site longitude (a numeric vector)}
    \item{\code{Setup_date}}{Date of deployment (a character vector)}
    \item{\code{Retrieval_date}}{Date of retrieval (a character vector)}
    \item{\code{DaysActive}}{How many days camera was functioning (a numeric vector)}
    \item{\code{DOY_1}}{Julian day 1 (a numeric vector)}
    \item{\code{DOY_2}}{Julian day 2 (a numeric vector)}
    \item{\code{DOY_3}}{Julian day 3 (a numeric vector)}
    \item{\code{DOY_4}}{Julian day 4 (a numeric vector)}
    \item{\code{DOY_5}}{Julian day 5 (a numeric vector)}
    \item{\code{DOY_6}}{Julian day 6 (a numeric vector)}
    \item{\code{DOY_7}}{Julian day 7 (a numeric vector)}
    \item{\code{DOY_8}}{Julian day 8 (a numeric vector)}
    \item{\code{DOY_9}}{Julian day 9 (a numeric vector)}
    \item{\code{DOY_10}}{Julian day 10 (a numeric vector)}
    \item{\code{DOY_11}}{Julian day 11 (a numeric vector)}
    \item{\code{DOY_12}}{Julian day 12 (a numeric vector)}
    \item{\code{DOY_13}}{Julian day 13 (a numeric vector)}
    \item{\code{DOY_14}}{Julian day 14 (a numeric vector)}
    \item{\code{DOY_15}}{Julian day 15 (a numeric vector)}
    \item{\code{DOY_16}}{Julian day 16 (a numeric vector)}
    \item{\code{DOY_17}}{Julian day 17 (a numeric vector)}
    \item{\code{DOY_18}}{Julian day 18 (a numeric vector)}
    \item{\code{DOY_19}}{Julian day 19 (a numeric vector)}
    \item{\code{DOY_20}}{Julian day 20 (a numeric vector)}
    \item{\code{DOY_21}}{Julian day 21 (a numeric vector)}
    \item{\code{DOY_22}}{Julian day 22 (a numeric vector)}
    \item{\code{DOY_23}}{Julian day 23 (a numeric vector)}
    \item{\code{DOY_24}}{Julian day 24 (a numeric vector)}
    \item{\code{DOY_25}}{Julian day 25 (a numeric vector)}
    \item{\code{DOY_26}}{Julian day 26 (a numeric vector)}
    \item{\code{DOY_27}}{Julian day 27 (a numeric vector)}
    \item{\code{DOY_28}}{Julian day 28 (a numeric vector)}
  }
}
\source{
URL: https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/ecy.3945
}
\references{
Palomo‐Munoz, G., T. J. Werdel, C. W. Piper, M. S. Peek, A. M. Ricketts, and A. A. Ahlers. 2023. Spatiotemporal distributions of mammals occurring in an agro‐prairie ecosystem. Ecology 104:e3945
}
\examples{
data(KScams_doy)
}
\keyword{datasets}
