% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_sumYear.R
\name{spp_sumYear}
\alias{spp_sumYear}
\title{Species Summary Table by Year}
\usage{
spp_sumYear(data, name_col, year_col)
}
\arguments{
\item{data}{The dataset containing information on species observations.}

\item{name_col}{The column name indicating the species in the dataset.}

\item{year_col}{The column name indicating the year in the dataset.}
}
\value{
A list where each element corresponds to a unique year and contains a data frame with three columns:
  \describe{
    \item{Name}{The unique species names.}
    \item{Observations}{The number of observations for each unique species in that year.}
  }
}
\description{
Function generates a list summarizing the number of observations for each unique species in a dataset by year.
}
\examples{
spp_sumYear_ex <- spp_sumYear(KScams_dat, name_col = "Common_name", year_col = "Year")

}
